+++ vivarium/core/serialize.py
@@ -1,3 +1,12 @@
+"""
+==========
+Serialize
+==========
+
+Collection of serializers that transform Python data into
+a BSON-compatible form.
+"""
+
 import re
 import math
 import warnings
@@ -18,6 +27,17 @@ def serialize_value(
     value: Any,
     codec_options: CodecOptions=None
 ) -> Any:
+    """Apply PyMongo's TypeCodec-based serialization routine on ``value``.
+
+    Args:
+        value (Any): Data to be serialized
+        codec_options (bson.codec_options.CodecOptions): Options used when
+            encoding / decoding BSON. Defaults to None, in which case options
+            are generated from the currently registered serializer codecs.
+
+    Returns:
+        Any: Serialized data
+    """
     if not codec_options:
         codec_options = get_codec_options()
     value = _dict_to_bson(value, False, codec_options)
@@ -27,6 +47,21 @@ def serialize_value(
 # BSON C extensions cannot distinguish between strings that
 # should be deserialized as different types (e.g. Units)
 def deserialize_value(value: Any) -> Any:
+    """Find and apply the correct serializer for a value
+    by calling each registered serializer's
+    :py:meth:`vivarium.core.registry.Serializer.can_deserialize()`
+    method. Returns the value as is if no compatible serializer
+    is found.
+
+    Args:
+        value (Any): Data to be deserialized
+
+    Raises:
+        ValueError: Only one serializer should apply for any given value
+
+    Returns:
+        Any: Deserialized data
+    """
     compatible_serializers = []
     for serializer_name in serializer_registry.list():
         serializer = serializer_registry.access(serializer_name)
@@ -44,7 +79,8 @@ def deserialize_value(value: Any) -> Any:
 
 
 class SequenceDeserializer(Serializer):  # pylint: disable=abstract-method
-
+    """Iterates through lists and applies deserializers.
+    """
     def can_deserialize(self, data: Any) -> bool:
         return isinstance(data, list)
 
@@ -53,7 +89,8 @@ class SequenceDeserializer(Serializer):  # pylint: disable=abstract-method
 
 
 class DictDeserializer(Serializer):  # pylint: disable=abstract-method
-
+    """Iterates through dictionaries and applies deserializers.
+    """
     def can_deserialize(self, data: Any) -> bool:
         return isinstance(data, dict)
 
@@ -65,7 +102,9 @@ class DictDeserializer(Serializer):  # pylint: disable=abstract-method
 
 
 class UnitsSerializer(Serializer):
-    """Serializer for data with units."""
+    """Serializes data with units into strings of the form ``!units[...]``,
+    where ``...`` is the result of calling ``str(data)``. Deserializes strings
+    of this form back into data with units."""
 
     def __init__(self) -> None:
         super().__init__(name='units')
@@ -107,7 +146,7 @@ class UnitsSerializer(Serializer):
 
         Args:
             data: The data to deserialize. Providing a list here is
-                deprecated. You should use :py:func:`deserialize_value`
+                deprecated. You should use ``deserialize_value``
                 instead, which uses a separate list deserializer.
             unit: The units to convert ``data`` to after deserializing.
                 If omitted, no conversion occurs. This option is
@@ -162,6 +201,7 @@ class NumpySerializer(Serializer):
         return [self.Codec()]
 
 class NumpyBoolSerializer(Serializer):
+    """Serializer for ``np.bool_`` objects."""
     class Codec(TypeEncoder):
         python_type = np.bool_
         def transform_python(self, value: np.bool_) -> bool:
@@ -171,6 +211,7 @@ class NumpyBoolSerializer(Serializer):
         return [self.Codec()]
 
 class NumpyInt64Serializer(Serializer):
+    """Serializer for ``np.int64`` objects."""
     class Codec(TypeEncoder):
         python_type = np.int64
         def transform_python(self, value: np.int64) -> int:
@@ -180,6 +221,7 @@ class NumpyInt64Serializer(Serializer):
         return [self.Codec()]
 
 class NumpyInt32Serializer(Serializer):
+    """Serializer for ``np.int32`` objects."""
     class Codec(TypeEncoder):
         python_type = np.int32
         def transform_python(self, value: np.int32) -> int:
@@ -189,6 +231,7 @@ class NumpyInt32Serializer(Serializer):
         return [self.Codec()]
 
 class NumpyFloat32Serializer(Serializer):
+    """Serializer for ``np.float32`` objects."""
     class Codec(TypeEncoder):
         python_type = np.float32
         def transform_python(self, value: np.float32) -> float:
@@ -198,6 +241,7 @@ class NumpyFloat32Serializer(Serializer):
         return [self.Codec()]
 
 class SetSerializer(Serializer):
+    """Serializer for set objects."""
     class Codec(TypeEncoder):
         python_type = set
         def transform_python(self, value: set) -> List:
@@ -207,6 +251,7 @@ class SetSerializer(Serializer):
         return [self.Codec()]
 
 class FunctionSerializer(Serializer):
+    """Serializer for function objects."""
     class Codec(TypeEncoder):
         python_type = type(deserialize_value)
         def transform_python(self, value: Callable) -> str:
@@ -229,6 +274,9 @@ class ProcessSerializer(Serializer):
 # TypeEncoders require their own TypeEncoders.
 # This includes Process, Composites, etc.
 def get_codec_options() -> CodecOptions:
+    """Returns a set of options used for serializing and
+    deserializing BSON by collecting codecs from all registered
+    serializers."""
     codecs = []
     for serializer in serializer_registry.registry.values():
         codecs += serializer.get_codecs()