+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v1.5.3
+* (#211) Fix a memory leak when removing parallel processes.
+* (#204) Implement a new serialization API based on PyMongo's 
+  `TypeCodec` interface.
+
 ## v1.5.2
 
 * (#207) Fix a bug in `Store.apply_update()` that caused failures when